requirements.txt
pandas
openpyxl
pdfplumber
python-docx
groq
instructor
pydantic
python-dotenv
tqdm